Intermediate checkin (working but still a work in progress)
