+++ tableauserverclient/models/__init__.py
@@ -1,6 +1,7 @@
 from .connection_item import ConnectionItem
 from .datasource_item import DatasourceItem
 from .exceptions import UnpopulatedPropertyError
+from .extract_refresh_task_item import ExtractRefreshTaskItem
 from .group_item import GroupItem
 from .interval_item import IntervalItem, DailyInterval, WeeklyInterval, MonthlyInterval, HourlyInterval
 from .pagination_item import PaginationItem
@@ -8,7 +9,6 @@ from .project_item import ProjectItem
 from .schedule_item import ScheduleItem
 from .site_item import SiteItem
 from .tableau_auth import TableauAuth
-from .task_item import TaskItem
 from .user_item import UserItem
 from .view_item import ViewItem
 from .workbook_item import WorkbookItem
