+++ tableauserverclient/models/exceptions.py
@@ -2,4 +2,4 @@ class UnpopulatedPropertyError(Exception):
     pass
 
 class ResponseBodyError(Exception):
-    passs
+    pass
