+++ tableauserverclient/server/endpoint/__init__.py
@@ -6,6 +6,7 @@ from .groups_endpoint import Groups
 from .projects_endpoint import Projects
 from .schedules_endpoint import Schedules
 from .sites_endpoint import Sites
+from .tasks_endpoint import Tasks
 from .users_endpoint import Users
 from .views_endpoint import Views
 from .workbooks_endpoint import Workbooks
